requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
mangum==0.19.0
openai==1.78.1
python-dotenv==1.1.0
